document facilitation refined
