+++ src/modules/restaurants.ts
@@ -1,6 +1,6 @@
 import * as uuid from "uuid/v1";
 import { Main } from "../main";
-import { DatabaseMethods } from "../database";
+import { DatabaseMethods } from "./database";
 
 export class Restaurants {
     main: Main;
